refactor: merge answers and marks into single route and adjust endpoints
