+++ api/__init__.py
@@ -2,7 +2,6 @@ from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
 from api.router import api_router
-from api.router.answers import answer_router
 from api.router.exam import exam_router
 from api.router.mark import marking_router
 from api.router.questions import questions_router
@@ -50,6 +49,5 @@ def create_application() -> FastAPI:
     rest_api.include_router(exam_router)
     rest_api.include_router(questions_router)
     rest_api.include_router(marking_router)
-    rest_api.include_router(answer_router)
 
     return rest_api
