+++ api/router/mark.py
@@ -2,16 +2,17 @@ from fastapi import APIRouter, Depends
 from sqlmodel import Session, select
 
 from api.dependencies import get_assessment_id, get_session
+from api.schemas.answer import Answer, AnswerRead
 from api.schemas.mark import (
     Mark,
     MarkRead,
 )
 
-marking_router = APIRouter(prefix="/marks", tags=["marking"])
+marking_router = APIRouter(prefix="/{student_username}", tags=["marking"])
 
 
 @marking_router.get(
-    "/{student_username}",
+    "/marks",
     response_model=list[MarkRead],
     summary="Retrieve user marks and feedback for exam questions",
     description="""
@@ -30,5 +31,29 @@ def get_marks_for_question(
             Mark.username == student_username,
         )
         .order_by(Mark.question, Mark.part, Mark.section)  # type: ignore
+    )
+    return session.exec(query).all()
+
+
+@marking_router.get(
+    "/answers",
+    response_model=list[AnswerRead],
+    summary="Retrieve student answers for all questions in exam",
+    description="""
+Retrieve the latest answers for all questions in exam
+""",
+)
+def get_answers_for_student(
+    student_username: str,
+    session: Session = Depends(get_session),
+    assessment_id: str = Depends(get_assessment_id),
+):
+    query = (
+        select(Answer)
+        .where(
+            Answer.exam_id == assessment_id,
+            Answer.username == student_username,
+        )
+        .order_by(Answer.question, Answer.part, Answer.section, Answer.task)  # type: ignore
     )
     return session.exec(query).all()
